finish a studen class method that update gpa into profile class;semesterTrancript inputCourseInfo method change to update its courseGradeMap

+++ Project/src/profile/Profile.java
@@ -12,8 +12,8 @@ public class Profile {
 	private String address;
 	private String country;
 	private char gender;
-	private double gpa;
-	private double currentSemesterGPA;
+	private float gpa;
+	private float currentSemesterGPA;
 	private String studentId;
